backend/requirements.txt
Django==5.1.4
djangorestframework
django-cors-headers
psycopg2-binary
asgiref==3.11.0
cffi==2.0.0
cryptography==46.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
psycopg2-binary==2.9.11
pycparser==3.0
sqlparse==0.5.5